refactor: separando responsabilidades do buttonManager
